--- ios/HackerNews/Components/StoryRowV2.swift
@@ -1,18 +1,0 @@
-//
-//  StoryRowV2.swift
-//  HackerNews
-//
-//  Created by Rikin Marfatia on 11/20/24.
-//
-
-import SwiftUI
-
-struct StoryRowV2: View {
-    @ObservedObject var model: AppViewModel
-    let story: Story
-    
-    var body: some View {
-      Text("Hello There")
-    }
-}
-
