+++ ios/HackerNews/Data/Bookmarks.swift
@@ -112,19 +112,3 @@ class LiveBookmarksDataStore: BookmarksDataStore {
     }
   }
 }
-
-struct FakeBookmarkDataStore: BookmarksDataStore {
-  func fetchBookmarks() -> [Bookmark] {
-    return []
-  }
-  
-  func addBookmark(_ bookmark: Bookmark) {
-  }
-
-  func removeBookmark(with id: Int64) {
-  }
-
-  func containsBookmark(with id: Int64) -> Bool {
-    return false
-  }
-}
